requirements.txt
streamlit
python-dotenv
pydantic
langchain
langchain-groq
langchain-community
pypdf
docx2txt
tenacity